+++ src/main/java/com/cardano/explorer/model/response/pool/projection/PoolDetailDelegatorProjection.java
@@ -1,6 +1,6 @@
 package com.cardano.explorer.model.response.pool.projection;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 public interface PoolDetailDelegatorProjection {
@@ -9,7 +9,7 @@ public interface PoolDetailDelegatorProjection {
 
   Timestamp getTime();
 
-  BigDecimal getFee();
+  BigInteger getFee();
 
   String getView();
 }
