+++ src/main/java/com/cardano/explorer/repository/AddressRepository.java
@@ -2,7 +2,7 @@ package com.cardano.explorer.repository;
 
 import com.sotatek.cardano.common.entity.Address;
 import com.sotatek.cardano.common.entity.StakeAddress;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
@@ -27,5 +27,5 @@ public interface AddressRepository extends JpaRepository<Address, Long> {
   Page<Address> findByStakeAddress(String stakeAddress, Pageable pageable);
 
   @Query("SELECT sum(addr.balance) FROM Address addr WHERE addr.stakeAddress = :stake")
-  Optional<BigDecimal> findTotalBalanceByStakeAddress(StakeAddress stake);
+  Optional<BigInteger> findTotalBalanceByStakeAddress(StakeAddress stake);
 }
